services/telegram-bot/requirements.txt
aiogram==3.18.0
httpx==0.28.1
python-dotenv==1.0.1
aiohttp-socks

